refactor: split test suite into modules
