+++ NightCityBot/tests/test_rolls.py
@@ -1,0 +1,23 @@
+from typing import List
+import discord
+from unittest.mock import AsyncMock, MagicMock, patch
+import config
+from NightCityBot.utils.constants import ROLE_COSTS_BUSINESS, ROLE_COSTS_HOUSING
+
+async def run(suite, ctx) -> List[str]:
+    """Test roll command functionality."""
+    logs = []
+    try:
+        logs.append("→ Expected: Valid roll should return a total, invalid roll should return a format error.")
+
+        roll_system = suite.bot.get_cog('RollSystem')
+        # Valid roll
+        await roll_system.loggable_roll(ctx.author, ctx.channel, "1d20+2")
+        logs.append("→ Result (Valid): ✅ Roll succeeded and result sent.")
+
+        # Invalid roll
+        await roll_system.loggable_roll(ctx.author, ctx.channel, "notadice")
+        logs.append("→ Result (Invalid): ✅ Error message shown for invalid format.")
+    except Exception as e:
+        logs.append(f"❌ Exception in test_rolls: {e}")
+    return logs
