+++ NightCityBot/tests/test_trauma_payment.py
@@ -1,0 +1,19 @@
+from typing import List
+import discord
+from unittest.mock import AsyncMock, MagicMock, patch
+import config
+from NightCityBot.utils.constants import ROLE_COSTS_BUSINESS, ROLE_COSTS_HOUSING
+
+async def run(suite, ctx) -> List[str]:
+    """Test Trauma Team subscription processing."""
+    logs = []
+    try:
+        user = await suite.get_test_user(ctx)
+        logs.append("→ Expected: collect_trauma should find thread and log subscription payment.")
+
+        economy = suite.bot.get_cog('Economy')
+        await economy.collect_trauma(ctx, user)
+        logs.append("→ Result: ✅ Trauma Team logic executed on live user (check #tt-plans-payment).")
+    except Exception as e:
+        logs.append(f"❌ Exception in test_trauma_payment: {e}")
+    return logs
